refactor: consolidate QMetry folder parameter interfaces by replacing test-case-folder interfaces with qmetry-test-case-folder interfaces
