+++ src/interfaces/index.ts
@@ -1,6 +1,3 @@
 export * from './list-folders-params.js';
-export * from './test-case-folder/list-projects-params.js';
-export * from './test-case-folder/create-folder-params.js';
-export * from './test-case-folder/update-folder-params.js';
-export * from './test-case-folder/copy-folder-params.js';
-export * from './test-case-folder/move-folder-params.js';
+export * from './qmetry-test-case-folder.js';
+export * from './qmetry-test-cycle-folders.js';
