--- src/interfaces/test-case-folder/copy-folder-params.ts
@@ -1,10 +1,0 @@
-/**
- * Parameters for copying QMetry test case folders
- */
-export interface CopyFolderParams {
-  folderId: string;
-  projectId: string;
-  newParentId: string;
-}
-
-export default CopyFolderParams;
