--- src/interfaces/test-case-folder/move-folder-params.ts
@@ -1,18 +1,0 @@
-/**
- * Parameters for moving a QMetry test case folder
- */
-export interface MoveFolderParams {
-  /** ID of the folder to move */
-  folderId: string;
-  
-  /** Project ID that contains the folder */
-  projectId: string;
-  
-  /** 
-   * ID of the new parent folder
-   * Use "-1" to move to the root level
-   */
-  newParentId: string;
-}
-
-export default MoveFolderParams;
